feat(CSSTransition): add "done" class for appear
This is not a breaking change because it doesn't remove the
`*-enter-done` class name after appearing is done, instead
`*-appear-done` is added in addition to that class, which provides you
with a bit more control in case you wanted to define a specific
animation only when an element first appears in the DOM, not every time
it enters.

Closes #383.
Closes #327.

Changes came from #383, but had to rebase to resolve conflicts and make
modifications informed by the discussion in #327.